fix : BridgeGame 내부 outputview 의존성 제거
- BridgeGame 내부 outputview 의존성 제거하며, 관련 로직을 GameController 생성하여 이동
- 그에 따른 테스트 코드로 리팩토링
