+++ src/main/java/bridge/view/OutputView.java
@@ -35,7 +35,7 @@ public class OutputView {
 	public void printResult(String result, Integer count) {
 		System.out.println(USER_FINAL_RESULT + result);
 		System.out.println(
-			USER_TRY_COUNT_PHRASE + String.format("%d", count)); // 여기서 string.format 이용해서 출력받기
+			USER_TRY_COUNT_PHRASE + String.format("%d", count));
 	}
 
 	public void printError(String errorMessage) {
