+++ src/test/java/bridge/service/BridgeGameTest.java
@@ -1,0 +1,14 @@
+package bridge.service;
+
+import org.junit.jupiter.api.Test;
+
+class BridgeGameTest {
+
+	@Test
+	void move() {
+	}
+
+	@Test
+	void retry() {
+	}
+}
