+++ src/test/java/bridge/service/GameServiceTest.java
@@ -7,8 +7,6 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
 
 import bridge.domain.Bridge;
 import bridge.domain.BridgeMaker;
@@ -19,10 +17,10 @@ import bridge.repository.UserBridgeRepository;
 class GameServiceTest {
 
 	private BridgeNumberGenerator bridgeNumberGenerator;
+	private GameService gameService;
 	private BridgeMaker bridgeMaker;
 	private Bridge bridge;
 	private UserBridgeRepository userBridgeRepository;
-	private GameService gameService;
 
 	@BeforeEach
 	void setUp() {
@@ -43,28 +41,12 @@ class GameServiceTest {
 	void makeBridge() {
 		//given
 		Integer bridgeSize = 4;
-		List<String> bridgeOption = List.of("U", "D");
 
 		//when
 		List<String> result = gameService.makeBridge(bridgeSize);
 
 		//then
 		result.stream().forEach(i -> Assertions.assertThat(i).containsAnyOf("U", "D"));
-	}
-
-	@DisplayName("checkValidSpace 테스트 - 유저가 건널수 있는 다리인지 확인하는 테스트")
-	@ParameterizedTest
-	@CsvSource(value = {"U, 0, true", "D, 1, true", "U, 2, true", "D, 3, false"})
-	void checkValidSpace(String userSpace, Integer currentSpace, boolean isValid) {
-		//given
-		List<String> bridgeTest = List.of("U", "D", "U", "U");
-
-		//when
-		bridge.initBridge(bridgeTest);
-		boolean checkValidSpace = gameService.checkValidSpace(userSpace, currentSpace);
-
-		//then
-		Assertions.assertThat(checkValidSpace).isEqualTo(isValid);
 	}
 
 	@DisplayName("saveUserCorrectSpace 에서 upperBridge 에 O 저장 확인 테스트")
@@ -108,7 +90,7 @@ class GameServiceTest {
 
 	@DisplayName("saveUserWrongSpace 에서 lowerBridge 에 X 저장 확인 테스트")
 	@Test
-	void getUserBridgeStatusDtoInLowerBridge() {
+	void saveUserWrongSpaceInLowerBridge() {
 		//given
 		String userLocation = "D";
 
@@ -119,18 +101,7 @@ class GameServiceTest {
 		Assertions.assertThat(userBridgeStatus).isEqualTo("[   ]" + "\n" + "[ X ]");
 	}
 
-	@DisplayName("userBridgeRepository 안에 데이터가 있는 상태에서 gameService.clearUserBridge() 할시 데이터가 남아 있는 확인 테스트")
 	@Test
-	void clearUserBridge() {
-		//given
-		userBridgeRepository.saveUserSpace("U", "X");
-		userBridgeRepository.saveUserSpace("D", "X");
-
-		//when
-		gameService.clearUserBridge();
-		String userBridgeStatus = userBridgeRepository.findUserBridgeStatus();
-
-		//then
-		Assertions.assertThat(userBridgeStatus).isEqualTo("[  ]" + "\n" + "[  ]");
+	void getUserBridgeStatus() {
 	}
 }
